services/mcp/requirements.txt
# Core dependencies
uvicorn==0.24.0
pydantic>=2.11.7

# Database
pymongo==4.6.0
qdrant-client>=1.15.1

# AI/ML - Use newer openai version compatible with httpx>=0.28.1
openai>=1.98.0

# HTTP and utilities
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.1.0

# MCP
fastmcp>=2.11.0